# src/main/resources/static/openapi/banking-openapi.yaml
openapi: 3.0.3
info:
  title: Brite Banking API
  version: 1.0.0
  description: |
    REST API for the `org.brite.banking` module of the bankingservices Spring Boot application:
    account registration, lookup, deposit/withdrawal, closing (single and bulk), paginated
    account-status search, bank office/ATM location search, bank statements, plus the module's async-notification and
    Resilience4j payment demo endpoints.

    ## Gateway behavior (applies to every endpoint in this document)

    Every path below sits behind two servlet filters (`org.brite.banking.gateway`), which run
    *before* the controllers:

    1. **`BusinessTransactionIdFilter`** generates a UUID per request and returns it in the
       `X-BTID` response header. The same id is stamped on every server log line for that
       request, so quote it when reporting a problem.
    2. **`BankingRateLimitFilter`** requires the customer-id header (default `X-Customer-Id`,
       configurable via `banking.rate-limit.customer-header-name`) and caps each customer at
       `banking.rate-limit.requests-per-day` requests/day (default 1000). A missing/blank
       header is rejected with `400`; an exhausted quota with `429`. Both are **plain-text**
       bodies written by the filter itself, not the JSON/text bodies produced by the
       controllers' exception handler. Every non-rejected response carries
       `X-RateLimit-Limit` and `X-RateLimit-Remaining`.

...
    authentication - the API has no auth layer.

...

...

...

...
servers:
...
security:
  - CustomerId: []
paths:
...
  /v1/api/staff/customers/{customerId}/login:
    post:
      tags: [Staff]
      operationId: createCustomerLogin
      summary: Create a customer login
      description: |
        Requires the `MANAGE_CUSTOMER_LOGINS` privilege (managers and up). Gives a customer a login: a username
        (`^[a-z0-9._-]{3,50}$`, stored lowercase) and an exactly-8-digit password, stored only as a BCrypt hash. A customer
        needs a login to use the account and portal endpoints, which require the customer token from
        `POST /v1/api/customers/login`. The login starts `ACTIVE`. The password is masked in request logs and never returned.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema: {type: integer, format: int64}
          example: 11
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example:
              username: alice.smith
              password: '13572468'
      responses:
        '201':
          description: The login was created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginStatusView'}
        '400':
          description: >-
            Missing `username`/`password` (JSON validation), a username or password in the wrong format, a username that is
            taken, or a customer who already has a login. Plain text.
          content:
            text/plain:
              schema: {type: string}
              examples:
                password:
                  value: Password must be exactly 8 digits
                hasLogin:
                  value: Customer 11 already has a login
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: No customer with that id (or the token's employee no longer exists). Plain text.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /v1/api/staff/login:
    post:
      tags: [Login]
      operationId: staffLogin
      summary: Employee login
      description: |
        Verifies the username (case-insensitive) and 8-digit password and returns a signed **JWT access token**
        (HS256, 30 minutes by default, no refresh token) plus who the caller is. **Every other staff endpoint requires
        this token** as `Authorization: Bearer <token>`; the endpoint verifies the signature and expiry and takes the
        acting employee from the token's `sub`. Responses are `Cache-Control: no-store`. The token carries `iss`, `sub`
        (employee number / customer id), `type` (`employee`/`customer`), `role` (employees), `jti`, `iat`, `exp`;
        no personal data.
        Unknown username and wrong password give the same `401`. After 5 wrong passwords the login is `LOCKED` for
        15 minutes (`423`); while locked, even the right password is refused. A right password on an `INACTIVE` or
        `SUSPENDED` login is `403` (or the employee isn't `ACTIVE` - on leave or terminated). The password is never logged or returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example:
              username: lucas.meyer
              password: '20260010'
      responses:
        '200':
          description: Credentials verified; token issued.
          headers:
            Cache-Control:
              schema: {type: string, example: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StaffLoginResponse'}
        '400':
          description: Missing `username` or `password` (Spring Boot default JSON error).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
        '401':
          description: Unknown username or wrong password. Plain text.
          content:
            text/plain:
              schema: {type: string}
              example: Invalid username or password
        '403':
          description: Right password but the login is `INACTIVE`/`SUSPENDED` (or the employee isn't `ACTIVE` - on leave or terminated). Plain text.
          content:
            text/plain:
              schema: {type: string}
        '423':
          description: The login is `LOCKED`. Plain text.
          content:
            text/plain:
              schema: {type: string}
              example: 'Too many failed login attempts; locked until 2026-10-01T09:15:00'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/customers/login:
    post:
      tags: [Login]
      operationId: customerLogin
      summary: Customer login
      description: |
        Verifies the username (case-insensitive) and 8-digit password and returns a signed **JWT access token**
        (HS256, 30 minutes by default, no refresh token) plus who the caller is. **Every customer account and
        portal endpoint requires this token** (`/v1/api/accounts/**` and `/bff/v1/portal/**`, except the two that create a
        customer), and a customer can only reach their own accounts; it is refused on staff endpoints (`403`). A customer
        needs a login first: a manager creates it with `POST /v1/api/staff/customers/{customerId}/login`. Responses are `Cache-Control: no-store`. The token carries `iss`, `sub`
        (employee number / customer id), `type` (`employee`/`customer`), `role` (employees), `jti`, `iat`, `exp`;
        no personal data.
        Unknown username and wrong password give the same `401`. After 5 wrong passwords the login is `LOCKED` for
        15 minutes (`423`); while locked, even the right password is refused. A right password on an `INACTIVE` or
        `SUSPENDED` login is `403`. The password is never logged or returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example:
              username: customer0005
              password: '20260005'
      responses:
        '200':
          description: Credentials verified; token issued.
          headers:
            Cache-Control:
              schema: {type: string, example: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomerLoginResponse'}
        '400':
          description: Missing `username` or `password` (Spring Boot default JSON error).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
        '401':
          description: Unknown username or wrong password. Plain text.
          content:
            text/plain:
              schema: {type: string}
              example: Invalid username or password
        '403':
          description: Right password but the login is `INACTIVE`/`SUSPENDED`. Plain text.
          content:
            text/plain:
              schema: {type: string}
        '423':
          description: The login is `LOCKED`. Plain text.
          content:
            text/plain:
              schema: {type: string}
              example: 'Too many failed login attempts; locked until 2026-10-01T09:15:00'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: |
        Customer identifier used solely for per-customer daily rate limiting - **not**
        authentication; any non-blank value is accepted. The header name is configurable
        (`banking.rate-limit.customer-header-name`) and the check can be turned off with
        `banking.rate-limit.enabled: false`.

    CustomerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The customer access token returned by `POST /v1/api/customers/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. Required on every `/v1/api/accounts/**` and `/bff/v1/portal/**` operation except
        the two that create a customer (`POST /v1/api/accounts/newaccount`, `POST /bff/v1/portal/accounts/open`). It proves who
        logged in; the customer's login is re-checked on every request, and the customer can only reach their own accounts.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The employee access token returned by `POST /v1/api/staff/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. It proves who logged in; permissions and status are re-checked from the
        database on every request, so a suspended login stops working immediately.
...
  responses:
...
    Unauthorized:
      description: >-
        No bearer token, or the token is invalid, expired or forged. Plain-text body; the response carries
        `WWW-Authenticate: Bearer`. Written by the staff authentication filter, so there is no JSON body.
      headers:
        WWW-Authenticate:
          schema: {type: string, example: Bearer}
      content:
        text/plain:
          schema: {type: string}
          examples:
            missing:
              value: "Authentication required: send 'Authorization: Bearer <token>' from POST /v1/api/staff/login"
            invalid:
              value: Invalid or expired token
...
    EmployeeForbidden:
      description: >-
        The acting employee is not ACTIVE, their login is not ACTIVE (or they have none), or their role lacks the
        required privilege. Plain-text body.
      content:
        text/plain:
          schema: {type: string}
          examples:
            privilege:
              value: 'Employee EMP-000010 (TELLER) is not authorized: requires SUSPEND_ACCOUNT'
            onLeave:
              value: Employee EMP-000016 is ON_LEAVE and cannot perform this action
            customerToken:
              value: This endpoint needs an employee token
            loginSuspended:
              value: 'Employee EMP-000010 login is SUSPENDED; only an ACTIVE login can perform transactions'
    TooManyRequests:
      description: >-
        The calling customer's daily quota (`banking.rate-limit.requests-per-day`) is
        exhausted. Written by the gateway filter as plain text. (A missing/blank customer-id
        header is rejected by the same filter with `400` and a plain-text message - see
        *Gateway behavior*.)
      headers:
        X-BTID:
          $ref: '#/components/headers/X-BTID'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        text/plain:
          schema: {type: string}
          example: 'Daily request limit exceeded for customer cust-42: max 1000 requests per day'
  headers:
    X-BTID:
      description: Business transaction id (UUID) for this request; appears on every server log line for it.
      schema: {type: string, format: uuid}
    X-RateLimit-Limit:
      description: Requests per day allowed for the calling customer.
      schema: {type: integer}
    X-RateLimit-Remaining:
      description: Requests remaining today for the calling customer.
      schema: {type: integer}
  schemas:
    LoginStatus:
      type: string
      enum: [ACTIVE, INACTIVE, LOCKED, SUSPENDED]
      description: |
        State of an employee or customer login; **only `ACTIVE` may perform transactions**. `LOCKED` is set
        automatically after too many wrong passwords (it expires at `lockedUntil`) or by an administrator
        (no expiry); `INACTIVE` and `SUSPENDED` are set by an administrator until set back to `ACTIVE`.
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, example: lucas.meyer}
        password: {type: string, description: Exactly 8 digits. Masked in request logs., example: '20260010'}
    StaffLoginResponse:
      type: object
      properties:
        accessToken: {type: string, description: Signed JWT (HS256). A credential - never log it., example: eyJhbGciOiJIUzI1NiJ9...}
        tokenType: {type: string, example: Bearer}
        expiresIn: {type: integer, format: int64, description: Seconds until the token expires., example: 1800}
        employee: {$ref: '#/components/schemas/Employee'}
    CustomerLoginResponse:
      type: object
      properties:
        accessToken: {type: string, description: Signed JWT (HS256). A credential - never log it., example: eyJhbGciOiJIUzI1NiJ9...}
        tokenType: {type: string, example: Bearer}
        expiresIn: {type: integer, format: int64, example: 1800}
        customer: {$ref: '#/components/schemas/AuthenticatedCustomer'}
    AuthenticatedCustomer:
      type: object
      description: Who a successful customer login belongs to; no date of birth, phone or address.
      properties:
        customerId: {type: integer, format: int64}
        firstName: {type: string}
        lastName: {type: string}
...
    LoginStatusView:
      type: object
      properties:
        username: {type: string}
        status: {$ref: '#/components/schemas/LoginStatus'}
        statusReason: {type: string, nullable: true}
        statusChangedAt: {type: string, format: date-time, nullable: true}
        lockedUntil: {type: string, format: date-time, nullable: true, description: Set only for an automatic lock.}
    EmployeeRole:
      type: string
      enum: [TELLER, MANAGER, AREA_MANAGER]
      description: |
        `TELLER`: VIEW_ACCOUNT, DEPOSIT, WITHDRAW. `MANAGER`: plus SUSPEND_ACCOUNT, UPDATE_SUSPENSION,
        REACTIVATE_ACCOUNT, CLOSE_ACCOUNT, VIEW_BRANCH_REPORTS, MANAGE_CUSTOMER_LOGINS. `AREA_MANAGER`: plus MANAGE_EMPLOYEES.
    EmployeePrivilege:
      type: string
      enum: [VIEW_ACCOUNT, DEPOSIT, WITHDRAW, SUSPEND_ACCOUNT, UPDATE_SUSPENSION, REACTIVATE_ACCOUNT, CLOSE_ACCOUNT, VIEW_BRANCH_REPORTS, MANAGE_EMPLOYEES, MANAGE_CUSTOMER_LOGINS]
    Employee:
      type: object
      properties:
        id: {type: integer, format: int64}
        employeeNumber: {type: string, example: EMP-000010}
        firstName: {type: string}
        lastName: {type: string}
        email: {type: string}
        phoneNumber: {type: string, example: 512-555-0210}
        role: {$ref: '#/components/schemas/EmployeeRole'}
        jobTitle: {type: string}
        status: {type: string, enum: [ACTIVE, ON_LEAVE, TERMINATED]}
        hireDate: {type: string, format: date}
        bankLocationId: {type: integer, format: int64, nullable: true, description: Branch; null for area managers.}
        region: {type: string, nullable: true, description: Area covered; set for area managers.}
        supervisorId: {type: integer, format: int64, nullable: true}
        privileges:
          type: array
          description: Derived from `role`.
          items: {$ref: '#/components/schemas/EmployeePrivilege'}
...
    SpringErrorResponse:
      type: object
      description: >-
        Spring Boot's default error body, returned when request validation fails or a
        parameter can't be parsed. It carries no field-level detail by default.
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer, example: 400}
        error: {type: string, example: Bad Request}
        path: {type: string, example: /brite/v1/api/accounts/newaccount}